+++ zia/services/dlpdictionaries/dlpdictionaries.go
@@ -56,6 +56,26 @@ type DlpDictionary struct {
 
 	// List of Indexed Document Match (IDM) profiles and their corresponding match accuracy for custom DLP dictionaries.
 	IDMProfileMatchAccuracy []IDMProfileMatchAccuracy `json:"idmProfileMatchAccuracyDetails"`
+
+	// Indicates whether to exclude documents that are a 100% match to already-indexed documents from triggering an Indexed Document Match (IDM) Dictionary.
+	IgnoreExactMatchIdmDict bool `json:"ignoreExactMatchIdmDict,omitempty"`
+
+	// A true value denotes that the specified Bank Identification Number (BIN) values are included in the Credit Cards dictionary. A false value denotes that the specified BIN values are excluded from the Credit Cards dictionary.
+	// Note: This field is applicable only to the predefined Credit Cards dictionary and its clones.
+	IncludeBinNumbers bool `json:"includeBinNumbers,omitempty"`
+
+	//The list of Bank Identification Number (BIN) values that are included or excluded from the Credit Cards dictionary. BIN values can be specified only for Diners Club, Mastercard, RuPay, and Visa cards. Up to 512 BIN values can be configured in a dictionary.
+	//Note: This field is applicable only to the predefined Credit Cards dictionary and its clones.
+	BinNumbers []int `json:"binNumbers,omitempty"`
+
+	// ID of the predefined dictionary (original source dictionary) that is used for cloning. This field is applicable only to cloned dictionaries. Only a limited set of identification-based predefined dictionaries (e.g., Credit Cards, Social Security Numbers, National Identification Numbers, etc.) can be cloned. Up to 4 clones can be created from a predefined dictionary.
+	DictTemplateId int `json:"dictTemplateId,omitempty"`
+
+	// This field is set to true if the dictionary is cloned from a predefined dictionary. Otherwise, it is set to false.
+	PredefinedClone bool `json:"predefinedClone,omitempty"`
+
+	// This value is set to true if proximity length and high confidence phrases are enabled for the DLP dictionary.
+	ProximityLengthEnabled bool `json:"proximityLengthEnabled,omitempty"`
 }
 
 type Phrases struct {